refactor context folder
