--- src/context/index.ts
@@ -1,1 +1,0 @@
-export { AuthProvider, useAuth } from './auth-context';
